+++ CHANGELOG.md
@@ -2,4 +2,10 @@
 
 ### Added
 
-Favicon to cv
+- Smart log deduplication - groups consecutive similar static asset requests
+- Favicon to cv
+
+### Fixed
+
+- CSS Rewrite
+- Reduced log noise from static asset requests (SVG, CSS, JS, etc.)